moved nextauth route to apps folder
