+++ app/api/auth/authOptions.ts
@@ -59,4 +59,4 @@ export const authOptions: NextAuthOptions = {
   },
 };
 
-export default NextAuth(authOptions);
+export default authOptions;
