+++ pages/context/AuthProvider.tsx
@@ -1,0 +1,12 @@
+'use client'
+import { SessionProvider } from "next-auth/react"
+
+export default function AuthProvider({children}:{
+    children : React.ReactNode
+}){
+    return (
+        <SessionProvider>
+            {children}
+        </SessionProvider>
+    )
+}
