+++ pages/create/[[...postIdArr]].tsx
@@ -3,7 +3,7 @@ import { ZodError } from "zod";
 import { useRouter } from "next/router";
 import React, { useState, useEffect, Fragment, useRef } from "react";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../api/auth/[...nextauth]";
+import { authOptions } from "../../app/api/auth/authOptions";
 import { useForm } from "react-hook-form";
 import CustomTextareaAutosize from "../../components/CustomTextareAutosize/CustomTextareaAutosize";
 import toast, { Toaster } from "react-hot-toast";
